fix: satisfy managed MCP architecture gates
