+++ scripts/docs-alignment-reviews.json
@@ -86,7 +86,17 @@
     {"reminder":"configure-agent-runtime: reviewer should compare English and Chinese counterparts","fingerprint":"a918175fa02f066f7fe469b850cc21dc37142c69bd6e5837b94f4dd46f398961","classification":"false-positive","reviewed_revision":"r6z-16-conversation-runtime-review-2026-07-25"},
     {"reminder":"configure-agent-runtime: reviewer should compare public facts with AGENT.CONTROL.TOOLS.001, AGENT.RUNTIME.ADAPTERS.001, AGENT.RUNTIME.PERMISSIONS.001, RUN.RESULT.001","fingerprint":"a918175fa02f066f7fe469b850cc21dc37142c69bd6e5837b94f4dd46f398961","classification":"false-positive","reviewed_revision":"r6z-16-conversation-runtime-review-2026-07-25"},
     {"reminder":"runtime-types: reviewer should compare English and Chinese counterparts","fingerprint":"41c8a98803a9a204c6e96ca9de0a1500f134fb2695c7fcb4235b3dd5ac111446","classification":"false-positive","reviewed_revision":"r6z-16-conversation-runtime-review-2026-07-25"},
-    {"reminder":"runtime-types: reviewer should compare public facts with AGENT.RUNTIME.ADAPTERS.001, AGENT.CONTROL.TOOLS.001, AGENT.RUNTIME.PERMISSIONS.001","fingerprint":"41c8a98803a9a204c6e96ca9de0a1500f134fb2695c7fcb4235b3dd5ac111446","classification":"false-positive","reviewed_revision":"r6z-16-conversation-runtime-review-2026-07-25"}
+    {"reminder":"runtime-types: reviewer should compare public facts with AGENT.RUNTIME.ADAPTERS.001, AGENT.CONTROL.TOOLS.001, AGENT.RUNTIME.PERMISSIONS.001","fingerprint":"41c8a98803a9a204c6e96ca9de0a1500f134fb2695c7fcb4235b3dd5ac111446","classification":"false-positive","reviewed_revision":"r6z-16-conversation-runtime-review-2026-07-25"},
+    {"reminder":"built-in-browser: reviewer should compare English and Chinese counterparts","fingerprint":"fcfe3c870399762ec6baf6edc63d9c8fcab620c6969ed514706c58f9c79c0c30","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"},
+    {"reminder":"built-in-browser: reviewer should compare public facts with AGENT.BROWSER.001, AGENT.RUNTIME.PERMISSIONS.001","fingerprint":"fcfe3c870399762ec6baf6edc63d9c8fcab620c6969ed514706c58f9c79c0c30","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"},
+    {"reminder":"skills: reviewer should compare English and Chinese counterparts","fingerprint":"2037a21d1be0419a6039edc1a33b0b6183f773118440b657e01fb4f3b9e8bd18","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"},
+    {"reminder":"skills: reviewer should compare public facts with AGENT.SKILLS.001, AGENT.SKILL.TELEMETRY.001, AGENT.RUNTIME.PERMISSIONS.001","fingerprint":"2037a21d1be0419a6039edc1a33b0b6183f773118440b657e01fb4f3b9e8bd18","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"},
+    {"reminder":"configure-agent-runtime: reviewer should compare English and Chinese counterparts","fingerprint":"c6a5b2476e88102ff4066c8f863e56bc0725f25f7a71ef578d8eb6bfb9057938","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"},
+    {"reminder":"configure-agent-runtime: reviewer should compare public facts with AGENT.CONTROL.TOOLS.001, AGENT.RUNTIME.ADAPTERS.001, AGENT.RUNTIME.PERMISSIONS.001, RUN.RESULT.001","fingerprint":"c6a5b2476e88102ff4066c8f863e56bc0725f25f7a71ef578d8eb6bfb9057938","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"},
+    {"reminder":"runtime-types: reviewer should compare English and Chinese counterparts","fingerprint":"cc2e85ae41ea8636ee83e9bc9da0062081014cb042135b2b6878ba43238c7959","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"},
+    {"reminder":"runtime-types: reviewer should compare public facts with AGENT.RUNTIME.ADAPTERS.001, AGENT.CONTROL.TOOLS.001, AGENT.RUNTIME.PERMISSIONS.001","fingerprint":"cc2e85ae41ea8636ee83e9bc9da0062081014cb042135b2b6878ba43238c7959","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"},
+    {"reminder":"permissions-and-platforms: reviewer should compare English and Chinese counterparts","fingerprint":"1ad1135c5db89f5981faaf3fa8e61a0a6bb13bd68ea2fc1e33bc67834d6a7827","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"},
+    {"reminder":"permissions-and-platforms: reviewer should compare public facts with AGENT.RUNTIME.PERMISSIONS.001, AGENT.BROWSER.001","fingerprint":"1ad1135c5db89f5981faaf3fa8e61a0a6bb13bd68ea2fc1e33bc67834d6a7827","classification":"false-positive","reviewed_revision":"managed-mcp-release-review-2026-07-25"}
   ],
   "allowed_classifications": ["fixed", "intentional", "false-positive"]
 }
